refactor: decompose useConversation into focused sub-hooks
Extract useAskFlow, useCommandExecution, and useScopeInfo from the monolithic useConversation hook. Move turn helpers (replaceTurn, removeLastThinkingTurn, askClassifiedToTurn) into conversationTurns.ts.
